refactored users schema from auth to own file
