+++ schemas/users.py
@@ -1,0 +1,10 @@
+from pydantic import BaseModel
+
+from typing import Optional
+
+
+class User(BaseModel):
+    username: str
+    email: Optional[str] = None
+    full_name: Optional[str] = None
+    disabled: Optional[bool] = None
